// api/model/fcCarPart.ts
/**
 * CarPartStore-API
 * API des CarPartStore-Backends
 *
 * The version of the OpenAPI document: 1.0
 * Contact: ata@hbt.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FcCarPartType } from './fcCarPartType';
import { FcCarMarke } from './fcCarMarke';


export interface FcCarPart { 
    id: number;
    name: string;
    carModel: string;
    carPartType: FcCarPartType;
    carMarke: FcCarMarke;
    barCode: string;
    quantity: number;
    quantityLimit: number;
    packageCount: number;
    rabatt: number;
    price: number;
    volumeInCube: number;
    totalWeightInKg: number;
}